rebased to merged PR#153 and moved datasets to a dataset package
