+++ docs/api_documentation.rst
@@ -21,11 +21,6 @@ The API Documentation for glotaran is automatically created from its docstrings.
     glotaran.analysis.fitting
     glotaran.analysis.grouping
     glotaran.analysis.simulation
-    glotaran.analysis.test
-    glotaran.analysis.test.mock
-    glotaran.analysis.test.test_fitting
-    glotaran.analysis.test.test_grouping
-    glotaran.analysis.test.test_simulation
     glotaran.analysis.variable_projection
     glotaran.dataio
     glotaran.dataio.chlorospec_format
@@ -35,13 +30,15 @@ The API Documentation for glotaran is automatically created from its docstrings.
     glotaran.dataio.legacy_readers
     glotaran.dataio.mapper
     glotaran.dataio.mlsd_file_format
-    glotaran.dataio.specialized_datasets
     glotaran.dataio.spectral_timetrace
     glotaran.dataio.wavelength_time_explicit_file
+    glotaran.datasets
+    glotaran.datasets.dataset
+    glotaran.datasets.specialized_datasets
+    glotaran.datasets.spectral_temporal_dataset
     glotaran.model
     glotaran.model.base_model
     glotaran.model.compartment_constraints
-    glotaran.model.dataset
     glotaran.model.dataset_descriptor
     glotaran.model.model
     glotaran.model.model_item
@@ -65,7 +62,6 @@ The API Documentation for glotaran is automatically created from its docstrings.
     glotaran.models.spectral_temporal.kinetic_result
     glotaran.models.spectral_temporal.spectral_matrix
     glotaran.models.spectral_temporal.spectral_shape
-    glotaran.models.spectral_temporal.spectral_temporal_dataset
     glotaran.models.spectral_temporal.spectral_temporal_dataset_descriptor
     glotaran.parse
     glotaran.parse.parser
