+++ glotaran/__init__.py
@@ -6,7 +6,7 @@
 #  from glotaran.specification_parser import parse_yml  # noqa: E402
 #  from glotaran.specification_parser import parse_file  # noqa: E402
 #  from glotaran import dataio, model  # noqa: E402
-from glotaran import model  # noqa: E402
+from . import model  # noqa: E402
 
 __version__ = '0.0.8'
